requirements.txt
streamlit
pandas
numpy
matplotlib
joblib
scikit-learn
torch
torchvision
Monai
gdown
pillow
scipy
